Подправлен стиль полей комментария

+++ media/admin.py
@@ -1,13 +1,9 @@
 import os
-from pathlib import Path
 
 from django.conf import settings
 from django.contrib import admin
 from django.utils.html import format_html
 from .models import Image
-import logging
-
-logger = logging.getLogger(__name__)
 
 class ImageAdmin(admin.ModelAdmin):
     list_display = ('category', 'image_filename', 'image_preview')
@@ -20,7 +16,8 @@ class ImageAdmin(admin.ModelAdmin):
             absolute_path = os.path.abspath(os.path.join(settings.MEDIA_ROOT, filename))
             # print(absolute_path)
             # print(filename)
-            os.remove(absolute_path)
+            if os.path.isfile(absolute_path):
+                os.remove(absolute_path)
             # obj.delete()
         return super().delete_queryset(request, queryset)
